+++ README.md
@@ -1,7 +1,10 @@
 # witness
 
 A minimal ETL library.
+
 Intended to be simple and concise.
+
+Made mostly for fun.
 
 ## Installation
 ```
