+++ requirements.txt
@@ -3,6 +3,7 @@
 requests>=2.0.0
 pyyaml>=6.0
 pyarrow>=8.0.0
+pendulum~=2.1.0
 
 # additional
 numpy~=1.23.0; python_version >= '3.9'
