+++ setup.cfg
@@ -30,6 +30,7 @@ install_requires =
     pandas>=1.4.0
     sqlalchemy>=1.4.0
     requests>=2.0.0
+    pendulum>=2.1.0
 
 
 [options.packages.find]
