+++ witness/extractors/database.py
@@ -25,9 +25,6 @@ class DatabaseExtractor(AbstractExtractor):
     def __init__(self, uri):
         super().__init__(uri)
 
-    def _set_extraction_timestamp(self):
-        setattr(self, 'extraction_timestamp', datetime.now())
-
     def extract(self):
         self._set_extraction_timestamp()
 
@@ -41,10 +38,6 @@ class ODBCExtractor(DatabaseExtractor):
         self.uri: str = uri
         self.query: str = query if isinstance(query, str) and is_select(query) else None
         super().__init__(uri)
-
-    def _set_extraction_timestamp(self):
-        from datetime import datetime
-        setattr(self, 'extraction_timestamp', datetime.now())
 
     def extract(self):
         from pyodbc import connect
